Create Celsius into Fahrenheit
Write a program that converts a temperature given in Celsius into Fahrenheit. Test the program by performing a compile-time initialization of 32 for the original Celsius value. Display the values number aligned. The original temperature should show no digits after the decimal. One position following the decimal should be printed for the converted value. Be sure to provide labels for both values. Go into your source code and change the initialization value to 0. Rerun the application. Select additional test values and rerun the application