Create convert an integer number of seconds into an equivalent number of hours, minutes, and seconds
Assignments Design an application using methods that convert an integer number of seconds into an equivalent number of hours, minutes, and seconds. Use methods for entering the initial seconds, performing the computations, and displaying the results. You should have separate methods for each computation. Results should be formatted and printed in a tabular display with the values number aligned.